app/requirements.dev.txt
Django==5.1.1
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django_filter==24.3
drf_spectacular==0.27.2
mypy==1.11.2
django-stubs==5.1.0
django-filter-stubs==0.1.3
djangorestframework-stubs==3.15.1
ruff==0.6.8
importanize==0.7.0